Fix: refactor and reorganize modules
