+++ src/pyosmeta/__init__.py
@@ -1,5 +1,6 @@
-from .contributors import PersonModel, ProcessContributors
-from .parse_issues import ProcessIssues, ReviewModel
+from .contributors import ProcessContributors
+from .models import PersonModel, ReviewModel
+from .parse_issues import ProcessIssues
 
 # Trick suggested by flake8 maintainer to ensure the imports above don't
 # get flagged as being "unused"
