+++ src/pyosmeta/cli/update_contributors.py
@@ -4,8 +4,9 @@ from datetime import datetime
 
 from pydantic import ValidationError
 
-from pyosmeta.contributors import PersonModel, ProcessContributors
+from pyosmeta.contributors import ProcessContributors
 from pyosmeta.file_io import create_paths, load_pickle, open_yml_file
+from pyosmeta.models import PersonModel
 
 # TODO - https://stackoverflow.com
 # /questions/55762673/how-to-parse-list-of-models-with-pydantic
