+++ src/pyosmeta/parse_issues.py
@@ -1,320 +1,12 @@
 import re
 import warnings
-from datetime import datetime
 
 import requests
 from dataclasses import dataclass
-from pydantic import (
-    AliasChoices,
-    BaseModel,
-    ConfigDict,
-    Field,
-    field_validator,
-)
-from typing import Any, Optional
+from typing import Any
 
-from pyosmeta.contributors import ProcessContributors, UrlValidatorMixin
-
-
-def clean_date(a_date: Optional[str]) -> str:
-    """Cleans up a datetime from github and returns a date string
-
-    In some cases the string is manually entered month-day-year and in
-    others it's a gh time stamp. finally sometimes it could be missing
-    or text. handle all of those cases with this validator.
-    """
-
-    if a_date is None or a_date == "missing":
-        return "missing"
-    else:
-        try:
-            return (
-                datetime.strptime(a_date, "%Y-%m-%dT%H:%M:%SZ")
-                .date()
-                .strftime("%Y-%m-%d")
-            )
-        except TypeError as t_error:
-            print("Oops - missing data. Setting date to missing", t_error)
-            return "missing"
-
-
-def clean_name(a_str: str) -> str:
-    """Helper to strip unwanted chars from text"""
-
-    unwanted = ["(", ")", "@"]
-    for char in unwanted:
-        a_str = a_str.replace(char, "")
-
-    return a_str.strip()
-
-
-def parse_user_names(username: str) -> dict:
-    """Parses authors, contributors, editors and usernames from
-    the requested issues.
-
-    Parameters
-    ----------
-    username : str
-        The line with username (see Notes).
-
-    Returns
-    -------
-    dict
-        ``{name: str, github_username: str}``
-
-    Notes
-    -----
-    Possible combinations:
-
-    1. Name Surname (@Github_handle)
-    2. (@Github_handle)
-    3. @Github_handle
-
-    """
-    names = username.split("@", 1)
-    names = [x for x in names if len(x) > 0]
-
-    if len(names) > 1:
-        parsed = {
-            "github_username": clean_name(names[1]),
-            "name": clean_name(names[0]),
-        }
-
-    else:
-        parsed = {
-            "github_username": clean_name(names[0]),
-            "name": "",
-        }
-
-    return parsed
-
-
-def clean_markdown(txt: str) -> str:
-    """
-    Remove Markdown characters from the beginning or end of a string.
-
-    Parameters
-    ----------
-    txt : str
-        The input string containing Markdown characters.
-
-    Returns
-    -------
-    str
-        The input string with Markdown characters removed from the beginning
-        and end.
-    """
-
-    pattern = r"^[`*]+|[`*]+$"
-
-    # Use re.sub to remove the matched Markdown characters
-    cleaned = re.sub(pattern, "", txt)
-
-    return cleaned
-
-
-class GhMeta(BaseModel, UrlValidatorMixin):
-    name: str
-    description: str
-    created_at: str
-    stargazers_count: int
-    watchers_count: int
-    open_issues_count: int
-    forks_count: int
-    documentation: Optional[str]  # Jointly is missing documentation
-    contrib_count: int
-    last_commit: str
-
-    @field_validator(
-        "last_commit",
-        "created_at",
-        mode="before",
-    )
-    @classmethod
-    def clean_date(cls, a_date: Optional[str]) -> str:
-        """Cleans up a datetime from github and returns a date string
-
-        Runs the general clean_date function in this module as a validator.
-        """
-
-        return clean_date(a_date)
-
-
-class ReviewModel(BaseModel):
-    # Make sure model populates both aliases and original attr name
-    model_config = ConfigDict(
-        populate_by_name=True,
-        str_strip_whitespace=True,
-        validate_assignment=True,
-    )
-
-    package_name: str | None = ""
-    package_description: str = Field(
-        "", validation_alias=AliasChoices("one-line_description_of_package")
-    )
-    submitting_author: dict[str, str | None] = {}
-    all_current_maintainers: list[dict[str, str | None]] = {}
-    repository_link: str | None = None
-    version_submitted: Optional[str] = None
-    categories: Optional[list[str]] = None
-    editor: dict[str, str | None] = {}
-    reviewer_1: dict[str, str | None] = {}
-    reviewer_2: dict[str, str | None] = {}
-    archive: str | None = None
-    version_accepted: str | None = None
-    date_accepted: str | None = Field(
-        default=None,
-        validation_alias=AliasChoices(
-            "date_accepted_(month/day/year)", "Date accepted"
-        ),
-    )
-    created_at: str = None
-    updated_at: str = None
-    closed_at: Optional[str] = None
-    issue_link: str = None
-    joss: Optional[str] = None
-    partners: Optional[list[str]] = None
-    gh_meta: Optional[GhMeta] = None
-
-    @field_validator(
-        "date_accepted",
-        mode="before",
-    )
-    @classmethod
-    def clean_date_review(cls, a_date: Optional[str]) -> str:
-        """Clean a manually added datetime that is added to a review by an
-        editor when the review package is accepted.
-
-        """
-        if a_date is None or a_date in ["missing", "TBD"]:
-            return "missing"
-        else:
-            new_date = a_date.replace("/", "-").split("-")
-            if len(new_date[0]) == 4:
-                return f"{new_date[0]}-{new_date[1]}-{new_date[2]}"
-            else:
-                return f"{new_date[2]}-{new_date[0]}-{new_date[1]}"
-
-    @field_validator(
-        "created_at",
-        "updated_at",
-        "closed_at",
-        mode="before",
-    )
-    @classmethod
-    def clean_date(cls, a_date: Optional[str]) -> str:
-        """Cleans up a datetime from github and returns a date string
-
-        Runs the general clean_date function in this module as a validator.
-
-        """
-
-        return clean_date(a_date)
-
-    @field_validator(
-        "package_name",
-        mode="before",
-    )
-    @classmethod
-    def clean_pkg_name(cls, pkg_name: str) -> str:
-        """A small cleaning step to remove any additional markdown
-        from a package's name
-
-        Parameters
-        ----------
-        pkg_name : str
-            Name of a pyOpenSci package extracted from review issue title
-
-        Returns
-        -------
-        str
-            Cleaned string with any markdown formatting removed.
-        """
-
-        return clean_markdown(pkg_name)
-
-    @field_validator(
-        "repository_link",
-        mode="before",
-    )
-    @classmethod
-    def clean_markdown_url(cls, repo: str) -> str:
-        """Remove markdown link remnants from gh usernames and name.
-
-        Sometimes editors and reviewers add names using github links.
-        Remove the link data.
-        """
-
-        if repo.startswith("["):
-            return repo.split("](")[0].replace("[", "")
-        else:
-            return repo
-
-    @field_validator(
-        "editor",
-        "reviewer_1",
-        "reviewer_2",
-        mode="before",
-    )
-    @classmethod
-    def clean_gh_url(cls, user: dict[str, str]) -> dict[str, str]:
-        """Remove markdown link remnants from gh usernames and name.
-
-        Sometimes editors and reviewers add names using github links.
-        Remove the link data.
-        """
-
-        user["github_username"] = user["github_username"].replace(
-            "https://github.com/", ""
-        )
-        user["name"] = re.sub(r"\[|\]", "", user["name"])
-
-        return user
-
-    @field_validator(
-        "categories",
-        mode="before",
-    )
-    @classmethod
-    def clean_categories(cls, categories: list[str]) -> list[str]:
-        """Make sure each category in the list is a valid value.
-
-        Valid pyos software categories are:
-            citation-management-bibliometrics, data-deposition,
-            data-extraction, data-processing-munging, data-retrieval,
-            data-validation-testing, data-visualization-analysis,
-            database-interoperability, education,
-            geospatial, scientific-software-wrappers,
-            workflow-automation-versioning
-
-
-        Parameters
-        ----------
-        categories : list[str]
-            List of categories to clean.
-
-        Returns
-        -------
-        list[str]
-            List of cleaned categories.
-        """
-
-        valid_categories = {
-            "data-processing": "data-processing-munging",
-            "data-validation": "data-validation-testing",
-            "scientific-software": "scientific-software-wrapper",
-        }
-
-        cleaned_cats = []
-        for category in categories:
-            for valid_prefix, valid_cat in valid_categories.items():
-                if category.startswith(valid_prefix):
-                    cleaned_cats.append(valid_cat)
-                    break
-            else:
-                # No match found, keep the original category
-                cleaned_cats.append(category)
-        return cleaned_cats
+from .clean import clean_markdown, parse_user_names
+from .contributors import ProcessContributors
 
 
 @dataclass
@@ -322,11 +14,9 @@ class ProcessIssues:
     """
     A class that processes GitHub issues in our peer review process and returns
     metadata about each package.
-
-
     """
 
-    def __init__(self, org, repo_name, label_name):
+    def __init__(self, org: str, repo_name: str, label_name: str):
         """
         More here...
